remove KeyId and marshal/unmarshal
