--- keys/keys_test.go
@@ -1,5 +1,0 @@
-package keys
-
-type TestMessage struct {
-	Hello string `json:"hello"`
-}
