+++ signature/rsa.go
@@ -1,4 +1,4 @@
-package keys
+package signature
 
 //import (
 //	"bytes"
